Fix rendering issues on workshop and social pages (#77)
This fixes a strange issue only see when the JS is minified for prod and where the user directly navigates to a hash route under the program page/component. Basically the html was not rendering as described in the component, p tags were closing before they should, and other things weren't right. This simply updates the html structure to fix whatever deep underlying issues in either react, the router, or webpack compilation, that might be contributing to the original behavior. Still don't understand why, but we won't look a gifted horse in the mouth on this one.